backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==2.10.0
python-dotenv==1.0.0
Pillow==10.2.0

